Move share details to separate file
