+++ node-src/lib/share/index.ts
@@ -1,2 +1,2 @@
 export { confirmShare, type ConfirmShareStatus } from './confirm';
-export { reserveShareOnAPI } from './upload';
+export { reserveShare } from './upload';
